Fix wave random at high level-times

+++ code/renderercommon/tr_common.h
@@ -123,7 +123,7 @@ extern	cvar_t	*r_saveFontData;
 qboolean	R_GetModeInfo( int *width, int *height, float *windowAspect, int mode );
 
 float R_NoiseGet4f( float x, float y, float z, double t );
-int R_RandomOn( float t );
+int R_RandomOn( double t );
 void  R_NoiseInit( void );
 
 image_t     *R_FindImageFile( const char *name, imgType_t type, imgFlags_t flags );
